pyproject.toml
[project]
name = "dizionaut"
version = "0.1.0"
description = "Telegram bot for word translation using MyMemory API"
authors = [
    {name = "ilbagatto",email = "sergei.krushinski@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "aiogram (>=3.20.0.post0,<4.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)"
]

[tool.poetry]
packages = [{include = "dizionaut", from = "src"}]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "-ra -q"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
